txscript: Optimize typeOfScript stakerev detection.
This continues the process of converting the typeOfScript function to
use a combination of raw script analysis and the new tokenizer instead
of the far less efficient parsed opcodes.

In particular, it converts the detection of stake revocation scripts to
use raw script analysis.

In order to accomplish this, it introduces a new function named
isStakeGenScript which makes of the recently added
extractStakePubKeyHash and extractStakeScriptHash functions.

The following is a before and after comparison of analyzing a large
script:

benchmark                          old ns/op    new ns/op    delta
---------------------------------------------------------------------
BenchmarkIsStakeRevocationScript   117699       4.58         -100.00%

benchmark                          old allocs   new allocs   delta
---------------------------------------------------------------------
BenchmarkIsStakeRevocationScript   1            0            -100.00%

benchmark                          old bytes    new bytes    delta
---------------------------------------------------------------------
BenchmarkIsStakeRevocationScript   466944       0            -100.00%