blockchain: Deprecate BlockOneCoinbasePaysTokens.
This deprecates the exported BlockOneCoinbasePaysTokens function in
favor of an unexported variant.  This is being done since it is only
intended to be an internal function and thus should not be exported.

However, since unexporting the function is a break to the API, it
requires a major version bump to the blockchain module.  Therefore, the
exported function is retained and simply redefined in terms of the now
unexported definition.