chaincfg: Explicit unique net addr prefix.
This modifies the registration process to enforce the already implicit
assumption that the network address prefix is unique per network so that
it is now an explicit requirement.

This is desirable because the address format contains a network
identification byte which is used elsewhere in the code to uniquely
identify which parameters to use.