chaincfg: Move mainnet code to mainnet files.
This moves the various code that is specific to mainnet to the
mainnetparams.go file to keep the information housed together.  Along
the same lines, it moves the tests specific to the mainnet params to
mainnetparams_test.go.

It also unexports the block on ledger definition since it should only be
exposed via the params.