chaincfg: Use scripts in block one token payouts.
This modifies the definitions for the required payouts in block one to
define required scripts instead of an address string as well as specify
the required script version and updates all of the network params
accordingly.

This is being done because addresses are a non-consensus construct that
essentially provide a human-readable method of specifying the actual
underlying cryptographic primitives.  Thus, the consensus rules
themselves must not directly rely on the ability to interpret and parse
addresses.

It is important to note that, once the blockchain code is updated to
make use of the changes this introduces, it is technically a fork since
the code will be more restrictive, but since it only applies to block
one, it's irrelevant because:

- It's buried under years of blocks
- The actual ledger of block one is unchanged