dcrutil: Reduce global refs in addr unit tests.
This modifies the test code for the address unit tests to prefer using
local references over accessing the global parameters directly from the
chaincfg package.  It is a good idea to avoid global references as much
as possible in general for several reasons, but the primary goal of this
change is to help reduce the tight coupling and therefore simplify
refactors.