txscript: Remove SigHashOptimization constant.
This removes the exported SigHashOptimization constant because it is not
a per-chain parameter and therefore does not belong here.

Further, it is only used in txscript and since it's an exported
constant, it couldn't be changed without a major version bump, which
would cause everything that relies on it to also bump majors.