txscript: Remove CPUMinerThreads constant.
This removes the exported CPUMinerThreads constant because it is not a
per-chain parameter and therefore does not belong here.

Further, it is only used in the mining code and since it's an exported
constant, it couldn't be changed without a major version bump and there
is no reason this shouldn't be defined in the mining code itself.