chaincfg: Remove all registration capabilities.
This removes all of the global registration capabilities from chaincfg
as part of the overall drive towards removing side effects that alter
the global state when importing packages and providing cleaner module
boundaries by reducing tight coupling betwee them.

For a bit of history, the package was designed back near the time Go
released before anyone had much experience with Go based on some other
early examples in the stdlib which many, including most of the original
authors, have come to realize is not ideal because it essentially alters
global state and has side effects merely by importing the package. This
can cause issues if multiple copies of the same package try to register
themselves.

Also, on a less theoretical note, with the benefit of hindsight and
experience, there isn't really any practical benefit to registering the
networks and allowing custom ones to be registered over just accepting a
params interface into the packages that need them whereas there are
tangible downsides having global registration factories as mentioned.
Accepting interfaces to the relevant functions in the other packages
will also significantly reduce the tight coupling that currently exists
with chaincfg and limit the scope of the needed parameters to
specifically those required by the functions in question.

An overview of the changes is as follows:

- Remove the following functions:
  - Register
  - IsPubKeyAddrID
  - IsPubKeyHashAddrID
  - IsPKHEdwardsAddrID
  - IsPKHSchnorrAddrID
  - IsScriptHashAddrID
  - HDPrivateKeyToPublicKeyID
  - ParamsByNetAddrPrefix
- Remove the following (now unused) errors:
  - ErrDuplicateNet
  - ErrUnknownHDKeyID
  - ErrDuplicateNetAddrPrefix
  - ErrUnknownNetAddrPrefix
- Remove all associated tests
- Update README to call out correct number of default net params