chaincfg; Add tests for required unique fields.
This adds a test to ensure that all of the network parameter fields that
are required to be unique are in fact unique for all of the provided
default parameters.

It should be noted that part of this is currently redundant with the
network registration code, however, that is going to be removed.