release: Freeze version 1 hdkeychain module.
This serves as the final release of version 1 of the hdkeychain module.
All future releases will be moving to version 2 of the module.

Since there have been no changes since the last release version, this
merely removes the hdkeychain override in the root module and updates it
so building the software will still produce binaries based on the v1
module until the v2 module is fully released.