hdkeychain: Remove Address method.
This removes the Address method from the ExtendedKey struct since
ultimately there will be multiple address types and therefore it no
longer makes sense to return an address which probably isn't the one the
caller actually needs.  Instead, the caller can make use of the ECPubKey
method to obtain the public key and create the desired address from
there.

It also removes the related tests, updates the example to show the
address conversion process for a standard pay-to-pubkey-hash address,
and updates the README.md and doc.go files accordingly.