hdkeychain: Remove SetNet method.
This removes the SetNet method from the ExtendedKey struct since its use
in practice could lead to subtle bugs due to the fact that changing the
network after keys have already been derived and then deriving more keys
will not result in the same derivation as starting with the target
network to begin with.  It is better to simply avoid the possibility of
misuse.

It also removes the related tests.