hdkeychain: Require network on decode extended key.
This modifies NewKeyFromString to accept the required network parameters
for the provided encoded extended key and introduces a new error to
indicate the provide extended key is for the wrong network.

This means that NewKeyFromString will now return ErrWrongNetwork if the
extended key being decoded is not for the provided network.  This
differs from the previous behavior relied on globally-registered
networks in chaincfg to determine which network the encoded key was for
followed by requiring the caller to check upon if the key was for the
specific network it desired upon return.

Finally, it removes the no longer necessary IsForNet method and updates
the documentation accordingly.