config: Cleanup and simplify network info parsing.
This simplifies and cleans up the initial network information parsing
which was recently added by replacing the networks map and all of the
unnecessary error handling that introduced with concrete supported
network definitions and directly updating their fields accordingly.

It also avoids a bunch of unnecessary conditional branching by setting
the flags to the results of the conditions directly.

Finally, it corrects the logic for determining if interface is limited
since that is currently only the case when the other interface has no
available listeners.