main: Convert signal handling to use context.
This converts the signal handling and dcrdMain function to use a context
with cancellation instead of a separate interrupt channel.  It does not
really make much of a difference yet, but the intent is to simplify the
various subsystems in future commits by refactoring them to make use of
the context versus manually handling start and stop in each subsystem.