miner: Improve background generator lifecycle.
This modifies the lifecycle of the background generator to use the
expected pattern for running subsystems based on contexts in the future.

In particular, this renames the Start function to Run and arranges for
it to block until the provided context is cancelled.  This is more
flexible for the caller since it can easily turn blocking code into
async code while the reverse is not true.

Next, the server is modified to ensure it waits for the background
template generator to shutdown before shutting down itself to help
ensure an orderly shutdown.

While here, it also makes the various lifecycle debug messages more
consistent with the rest of the code base.