mining: add NewBlockTemplateV2.
This adds modified versions (V2) of handleTooFewVoters
and NewBlockTemplate that take the current template as a parameter and updates call sites for the background
template generator accordingly.

An additional check has been added to prevent scheduling an
unnecessary regen if the side chain tip voted on has max votes required.