dcrutil: Require network on WIF decode.
This modifies DecodeWIF to accept the required network parameters for
the provided encoded WIF string and introduces a new error to indicate
the provided WIF string is for the wrong network.

This means that DecodeWIF will now return ErrWrongWIFNetwork if the WIF
string being decoded is not for the provided network.  This differs from
the previous behavior that would accept arbirtary network bytes and
relied on globally-registered networks in chaincfg to determine which
network the string was for followed by requiring the caller to check if
the WIF was for the specific network it desired upon return.

Finally, it removes the no longer necessary IsForNet method and updates
the documentation accordingly.