fullblocktests: Add coinbase nulldata tests.
This adds two new consensus tests to the full block tests for the
coinbase nulldata height commitment.  The first new test ensures
coinbase transactions with the wrong script version are rejected and the
second test ensures data pushes that exceed the maximum allowed number
of bytes for the coinbase nulldata push are rejected.