TravisCI: Test and lint latest version modules.
The current approach of the CI tests of specifying the path to the
directory that contains the module only tests version 1 of the module
despite there being newer version available.  The same is true for the
lints.

This resolves the test portion by obtaining a list of all modules
required by the root go.mod file and providing the entire module path,
which includes the version, to the go test command instead of the
directory.

Unfortunately, the linter does not recognize full module paths, however
it will lint the latest module version from within the module directory
itself, so this resolves the linting case by stripping the full module
path and version down to the directory and changing into the directory
to perform the lint checks.